Refactor file handling and validation logic in `DropzonePage`
- Adjusted file validators to align with the `accept` property for more precise validation.
- Avoided redundant gestures by adding early `return` in `GestureHint`.
- Simplified `fileInput` interaction and improved error toasts for storage failures.
- Updated CSS to prevent touch gesture interception in `PDFViewer`.

+++ src/lib/utils/imageImport.ts
@@ -72,7 +72,7 @@ async function webpToPng(webpBytes: Uint8Array): Promise<Uint8Array> {
             canvas.toBlob((b) => (b ? resolve(b) : reject(new Error('Canvas toBlob failed'))), 'image/png');
         });
 
-        return new Uint8Array(await pngBlob.arrayBuffer()) as Uint8Array;
+        return new Uint8Array(await pngBlob.arrayBuffer());
     } finally {
         URL.revokeObjectURL(url);
     }
